DRAFT: Redact environment variable not in policy
This is a DRAFT meant to generate discussion of the appropriate direction
for this proposed feature.

The MutateEnvVar flag when set to true is an enforcement mode that may change
incoming container group admissions to comply with the security policy.
For example, when MutateEnvVar is true container group environment
variables that are not included in the policy are redacted. If this flag
is false, the unexpected environment variable would block the container's
admission to the group.

Let's consider a concrete example. The OCI specification for the container
group environment variable section includes "CONTAINER_DEBUG=true". When
pkg/securitypolicy.enforceEnvironmentVariablePolicy determines that
there is no matching rule found for some container,
var:matchingRuleFoundForSomeContainer, and the MutateEnvVar flag is true,
the environment variable in the OCI specification will be redacted

as such:

"CONTAINER_DEBUG=true"

will become:

"_REDACTED_BY_POLICY_rand314159=CONTAINER_DEBUG=true"

The final redaction scheme is a work-in-progress. The current scheme will help with troubleshooting and testing.

Signed-off-by: Douglas MacIver <62668331+douglasmaciver@users.noreply.github.com>